little update inside start_gomoku.sh

+++ start_gomoku.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 DYLIB=target/debug/librust_ext.dylib
+LIBCURR=librust_ext.dylib
 PYTHON=src/PyInterface/gomoku.py
 maturin develop --release
-if test -f "$DYLIB"; then
-    mv $DYLIB .
+if test -f "$LIBCURR" && test -f "$DYLIB" ; then
+	rm $LIBCURR
+	mv $DYLIB .
 fi
-bin/python3.9 $PYTHON
+#if test -f "$DYLIB"; then
+    #mv $DYLIB .
+#fi
+venv/bin/python3 $PYTHON
 #maturin develop --release && mv target/debug/librust_ext.dylib . && bin/python3.9 src/PyInterface/gomoku.py
